half2

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -9,7 +9,7 @@
  */
 void puts_half(char *str)
 {
-	int i;
+	unsigned long int i;
 
 	if (strlen(str) % 2 == 0)
 	{
